fix: comment 수정, 삭제 작성자도 가능하도록 수정

+++ src/main/java/com/ktb10/munggaebe/post/service/CommentService.java
@@ -12,6 +12,7 @@ import com.ktb10.munggaebe.post.exception.CommentNotFoundException;
 import com.ktb10.munggaebe.post.exception.PostNotFoundException;
 import com.ktb10.munggaebe.post.repository.CommentRepository;
 import com.ktb10.munggaebe.post.repository.PostRepository;
+import com.ktb10.munggaebe.utils.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -91,13 +92,13 @@ public class CommentService {
 
 
     @Transactional
-    public Comment updateComment(final CommentServiceDto.UpdateReq updateReq, final long memberId) {
+    public Comment updateComment(final CommentServiceDto.UpdateReq updateReq) {
 
         final Long commentId = updateReq.getCommentId();
         final Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommentNotFoundException(commentId));
 
-        validateAuth(memberId, comment.getMember().getId());
+        validateAuthorization(comment);
 
         comment.updateComment(updateReq.getContent());
 
@@ -105,22 +106,20 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteComment(final long commentId, final long memberId) {
+    public void deleteComment(final long commentId) {
 
         final Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CommentNotFoundException(commentId));
 
-        validateAuth(memberId, comment.getMember().getId());
+        validateAuthorization(comment);
 
         comment.deleteComment();
     }
 
-    private void validateAuth(final long memberId, final long commentMemberId) {
-        final Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberNotFoundException(memberId));
-
-        if (member.getRole() == MemberRole.STUDENT && commentMemberId != memberId) {
-            throw new MemberPermissionDeniedException(memberId, member.getRole());
+    private void validateAuthorization(Comment comment) {
+        Long currentUserId = SecurityUtil.getCurrentUserId();
+        if (SecurityUtil.hasRole("STUDENT") && !comment.getMember().getId().equals(currentUserId)) {
+            throw new MemberPermissionDeniedException(currentUserId, MemberRole.STUDENT);
         }
     }
 }
